🚚 Extract DynamicMarkdownHeaderStyling.ts
